Fixed compiler warning on Lobo
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@2393 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/connections/RandomConn.cpp
@@ -50,7 +50,7 @@ int RandomConn::initializeUniformWeights(unsigned long seed)
    PVParams * params = parent->parameters();
 
    wMin = (long) params->value(name, "wMin", 0.0f);
-   idum = params->value(name, "idum", -1.0f);
+   idum = (long) params->value(name, "idum", -1.0f);
 
    float wMinInit = params->value(name, "wMinInit", wMin);
    float wMaxInit = params->value(name, "wMaxInit", wMax);
